fastwalk: only check if MSYSTEM is set during MSYS/MSYS2
This changes the MSYS/MSYS2 detection to only check if the MSYSTEM
environment variable is set instead of asserting on its value. This
is to support other MSYS environments like UCRT64.

It also removes the check for Windows Subsystem for Linux (WSL) since
it was unreliable and required us to stat and potentially read a file
on initialization (which is bad).

See:
 * https://github.com/charlievieth/fastwalk/commit/87029d9318159fb714bd064bbda22d69a11fa9f2#r144282929
 * https://github.com/sharkdp/fd/commit/f875ea9a5