fastwalk: don't clean the path argument to Walk on Windows
On Windows don't attempt to clean the path argument to Walk since the
existing clean logic (cleanRootPath) transforms paths like "C:\" => "C:"
which are not equivalent.

This logic only existed to make the joining of paths simpler and should
probably be removed since we shouldn't be modifying user provided paths.

TODO: Investigate if anything relies on the current clean logic and
remove it if nothing does.

Fixes: https://github.com/charlievieth/fastwalk/issues/37