darwin/arm64: use the number of perf cores as the default number of workers
Starting around macOS v15 the parallel performance of the APFS file
system appears slightly improved and now the ideal number of fastwalk
workers is generally the number of performance cores currently
available.

Therefore, this commit changes the default number of workers on
darwin/arm64 to be the value of sysctl("hw.perflevel0.physicalcpu")
which is the number of currently available performance cores (link to
xnu source below).

https://github.com/apple-oss-distributions/xnu/blob/43a90889846e00bfb5cf1d255cdc0a701a1e05a4/bsd/sys/sysctl.h#L1244